refactor(cli): move command bodies under cli/commands and keep main dispatcher thin
Made-with: Cursor
